kvserver: mv log compaction/size funcs to logstore
This commit moves the raft log storage helpers to the logstore package.

Epic: none
Release note: none
